LayoutTests/fast/dom/Window/Location/location-override-toString-using-defineGetter-expected.txt

PASS location.__defineGetter__('toString', getter) threw exception TypeError: Attempting to change configurable attribute of unconfigurable property..
PASS successfullyParsed is true

TEST COMPLETE
